+++ src/core/project.js
@@ -1,7 +1,7 @@
 // @flow
 
 import base64url from "base64url";
-import {type RepoId} from "../core/repoId";
+import {type RepoId} from "../plugins/github/repoId";
 import {toCompat, fromCompat, type Compatible} from "../util/compat";
 import {type Identity} from "../plugins/identity/identity";
 import {type DiscourseServer} from "../plugins/discourse/loadDiscourse";
